fix: 修复批量重构导致的多页面白屏并提交历史改动
- 修复 17 个后端页面因 $N 占位符残留与乱码引号丢失导致的 JS 语法崩溃/白屏
- 恢复或修复: 数据概览、AI助手、桌台扫码点餐、排队叫号、导出中心、菜品列表、食材预警、
  日营业报表、会员卡券详情/临期/发放/列表、会员留存分析、急单管理、智能补货、门店工作台/列表
- 全部 78 个后端页面内联 JS 通过语法校验, 全库无 $N 占位符残留
- 附带提交此前未提交的后端实体/配置/平台同步等改动
- 移除不再跟踪的 tests/ .github/ .mvn/ AGENTS.md (由 .gitignore 接管)

+++ src/main/java/com/reggie/module/ai/service/impl/UserProfileServiceImpl.java
@@ -412,14 +412,14 @@ public class UserProfileServiceImpl extends ServiceImpl<UserProfileMapper, UserP
         LambdaQueryWrapper<AIMessageRecord> wrapper = new LambdaQueryWrapper<>();
         wrapper.eq(AIMessageRecord::getUserId, userId)
                 .eq(AIMessageRecord::getRole, "user");
-        return Math.max(aiMessageRecordMapper.selectCount(wrapper) / 2, 0);
+        return Math.max(aiMessageRecordMapper.selectCount(wrapper).intValue() / 2, 0);
     }
 
     private int countUserFeedbacks(Long userId) {
         LambdaQueryWrapper<AIMessageRecord> wrapper = new LambdaQueryWrapper<>();
         wrapper.eq(AIMessageRecord::getUserId, userId)
                 .isNotNull(AIMessageRecord::getFeedback);
-        return (int) aiMessageRecordMapper.selectCount(wrapper);
+        return aiMessageRecordMapper.selectCount(wrapper).intValue();
     }
 
     private BigDecimal calculateConfidence(UserProfile profile) {
